Create README.md
Sepsis is a severe and potentially life-threatening condition that occurs when the body’s response to infection causes widespread inflammation, which can cause organs to malfunction and eventually lead to organ failure, such as acute kidney injury (AKI). When the body experiences an infection, the immune system releases chemicals to combat it, and these chemicals trigger widespread inflammation in sepsis. If we could detect these chemicals, also known as biomarkers, at the early stage of sepsis and provide prompt medical intervention or treatment, severe or even fatal conditions could be prevented. 
In this project, we analyzed a set of biomarkers data collected from the urine/blood output of rodent samples, which were either induced with bacterial or buffer solution, then we applied machine learning methods to identify the key biomarkers in the sepsis setting and evaluated the performance of these selected biomarkers.